Split up SimpleMode vs NormalMode
